Allow users set if to show close button or not
I built a game where a user had to click start before the game shows up. This is important as there is a multi-player feature. Now, I have a close button but I did not assign a function to it, which is really bad UX.
This prompted me to make this suggestion.